blender tutorial & SHanpe moving
We learnt blender again and it is so many. just learn what we need instead.

also 3D version of Move character is way more different. we are learning milestone!
we need to know how to roll based on what camera sees ahead and slide based on what camera sees.
we need contraining deadzone. from the off the deadzone, it must be zero. We thought constraint. So benefiting many controller to add more velocity likely be removed here. deadzone is min and 1 is max I think. ahh.